+++ tests/test_system.py
@@ -1,6 +1,10 @@
 import pytest
 
-from aiobsidian._exceptions import APIStatusError, AuthenticationError
+from aiobsidian._exceptions import (
+    APIParseError,
+    APIStatusError,
+    AuthenticationError,
+)
 from aiobsidian.models.system import ServerStatus
 
 STATUS_JSON = {
@@ -32,6 +36,13 @@ async def test_openapi(mock_api, client):
     result = await client.system.openapi()
 
     assert result == "openapi: 3.0.2"
+
+
+async def test_status_that_is_no_record(mock_api, client):
+    mock_api.get("/").respond(200, json=["OK"])
+
+    with pytest.raises(APIParseError):
+        await client.system.status()
 
 
 async def test_status_authentication_error(mock_api, client):